+++ engine/language-server/src/test/scala/org/enso/languageserver/search/SuggestionsHandlerSpec.scala
@@ -762,10 +762,10 @@ class SuggestionsHandlerSpec
               inserted(0),
               inserted(1),
               inserted(2),
-              inserted(6),
               inserted(7),
               inserted(8),
-              inserted(3)
+              inserted(9),
+              inserted(4)
             )
           )
         )
@@ -773,7 +773,7 @@ class SuggestionsHandlerSpec
 
     "search entries by self type" taggedAs Retry in withDb {
       (config, repo, _, _, handler) =>
-        val (_, Seq(_, _, _, methodId, _, _, methodOnAnyId, _, _)) =
+        val (_, Seq(_, _, _, getterId, methodId, _, _, methodOnAnyId, _, _)) =
           Await.result(repo.insertAll(Suggestions.all), Timeout)
         handler ! SearchProtocol.Completion(
           file       = mkModulePath(config, "Main.enso"),
@@ -787,7 +787,7 @@ class SuggestionsHandlerSpec
         expectMsg(
           SearchProtocol.CompletionResult(
             1L,
-            Seq(methodId, methodOnAnyId)
+            Seq(getterId, methodId, methodOnAnyId)
           )
         )
     }
@@ -796,7 +796,7 @@ class SuggestionsHandlerSpec
       (config, repo, _, _, handler) =>
         val (
           _,
-          Seq(_, _, _, _, _, _, anyMethodId, numberMethodId, integerMethodId)
+          Seq(_, _, _, _, _, _, _, anyMethodId, numberMethodId, integerMethodId)
         ) =
           Await.result(repo.insertAll(Suggestions.all), Timeout)
 
@@ -819,7 +819,7 @@ class SuggestionsHandlerSpec
 
     "search entries for any" taggedAs Retry in withDb {
       (config, repo, _, _, handler) =>
-        val (_, Seq(_, _, _, _, _, _, anyMethodId, _, _)) =
+        val (_, Seq(_, _, _, _, _, _, _, anyMethodId, _, _)) =
           Await.result(repo.insertAll(Suggestions.all), Timeout)
 
         handler ! SearchProtocol.Completion(
@@ -841,7 +841,7 @@ class SuggestionsHandlerSpec
 
     "search entries by return type" taggedAs Retry in withDb {
       (config, repo, _, _, handler) =>
-        val (_, Seq(_, _, _, _, functionId, _, _, _, _)) =
+        val (_, Seq(_, _, _, _, _, functionId, _, _, _, _)) =
           Await.result(repo.insertAll(Suggestions.all), Timeout)
         handler ! SearchProtocol.Completion(
           file       = mkModulePath(config, "Main.enso"),
@@ -862,7 +862,7 @@ class SuggestionsHandlerSpec
 
     "search entries by tags" taggedAs Retry in withDb {
       (config, repo, _, _, handler) =>
-        val (_, Seq(_, _, _, _, _, localId, _, _, _)) =
+        val (_, Seq(_, _, _, _, _, _, localId, _, _, _)) =
           Await.result(repo.insertAll(Suggestions.all), Timeout)
         handler ! SearchProtocol.Completion(
           file       = mkModulePath(config, "Main.enso"),
@@ -1089,7 +1089,7 @@ class SuggestionsHandlerSpec
       )
 
     val method: Suggestion.Method =
-      Suggestion.Method(
+      Suggestion.DefinedMethod(
         externalId    = Some(UUID.randomUUID()),
         module        = "Test.Main",
         name          = "main",