+++ lib/rust/ensogl/core/src/display/garbage.rs
@@ -18,6 +18,14 @@ struct Garbage {
     before_mouse_events: Vec<Box<dyn Any>>,
 }
 
+impl Garbage {
+    fn collected_items_count(&self) -> usize {
+        self.before_pixel_sync.len()
+            + self.before_pixel_update.len()
+            + self.before_mouse_events.len()
+    }
+}
+
 
 /// The Garbage Collector
 ///
@@ -89,6 +97,14 @@ impl Collector {
             std::mem::take(&mut garbage.before_mouse_events)
         };
         drop(before_mouse_events);
+    }
+
+    /// Immediately drop all collected garbage.
+    pub fn force_garbage_drop(&self) {
+        // Elements may add new objects on drop, thus we need to clear garbage in loop.
+        while self.garbage.borrow().collected_items_count() > 0 {
+            self.garbage.take();
+        }
     }
 }